Trait ink_storage_codegen::DeriveUtils

pub trait DeriveUtils {
    fn find_salt(&self) -> Option<TypeParam>;
    fn all_types(&self) -> Vec<Type>;
}

Provides common methods for DeriveInput.

Developer Note

This is only for internal usage in the codegen module.

Required Methods

fn find_salt(&self) -> Option<TypeParam>

Finds the salt of the structure, enum or union. The salt is any generic that has bound StorageKey. In most cases it is parent storage key or auto-generated storage key.

fn all_types(&self) -> Vec<Type>

Return all types of the input.